Move Menu class to separate file
